requirements.txt
# Flask & Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# API Documentation
Flasgger==0.9.7.1

# Blockchain Integration
web3==6.11.3

# Rate Limiting
Flask-Limiter==3.5.0

# Database (Optional - for PostgreSQL support)
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Security
cryptography==41.0.7
